Started working on Directory
Started working on Directory, but I think I’ll leave it for now. Not
really sure about what I want to do or how to do it, and not really
sure I need this class for now anyway. So I’ll commit what I have done
and then just ignore the files for now.